separating generating all data from all objects into separate function called GetAllObjects()
